+++ scripts/main.js
@@ -6,11 +6,45 @@ export const ITEM_ID = 'LUSjG8364p7LFY1u';
 
 export let dndV3 = false;
 
-CONFIG.debug.hooks = false;
+CONFIG.debug.hooks = true;
+
+(function () {
+  function checkCondition(v1, operator, v2) {
+    switch (operator) {
+      case '==':
+        return (v1 == v2);
+      case '===':
+        return (v1 === v2);
+      case '!==':
+        return (v1 !== v2);
+      case '<':
+        return (v1 < v2);
+      case '<=':
+        return (v1 <= v2);
+      case '>':
+        return (v1 > v2);
+      case '>=':
+        return (v1 >= v2);
+      case '&&':
+        return (v1 && v2);
+      case '||':
+        return (v1 || v2);
+      default:
+        return false;
+    }
+  }
+
+  Handlebars.registerHelper('ifCond', function (v1, operator, v2, options) {
+    return checkCondition(v1, operator, v2)
+      ? options.fn(this)
+      : options.inverse(this);
+  });
+}());
 
 Handlebars.registerHelper("spFormat", (path, ...args) => {
   return game.i18n.format(path, args[0].hash);
 });
+
 
 Hooks.on('init', () => {
   console.log('SpellPoints init');
@@ -54,6 +88,8 @@ Hooks.on('init', () => {
     }
   }
 
+  SpellPoints.setSpColors();
+
 });
 
 /** spell launch dialog **/
@@ -68,9 +104,19 @@ Hooks.on("preUpdateItem", SpellPoints.checkSpellPointsValues);
 
 //Hooks.on("dnd5e.computeLeveledProgression", SpellPoints.calculateSpellPointsProgression);
 
-Hooks.on("renderActorSheet5e", (app, html, data) => {
-  SpellPoints.alterCharacterSheet(app, html, data);
+Hooks.on("renderActorSheet5eCharacter2", (app, html, data) => {
+  SpellPoints.alterCharacterSheet(app, html, data, 'v2');
 });
+
+Hooks.on("renderActorSheet5eCharacter", (app, html, data) => {
+  SpellPoints.alterCharacterSheet(app, html, data, 'v1');
+});
+
+Hooks.on("renderActorSheet5eNPC", (app, html, data) => {
+  SpellPoints.alterCharacterSheet(app, html, data, 'npc');
+});
+
+
 
 /**
   * Hook that is triggered after the SpellPointsForm has been rendered. This
@@ -84,4 +130,8 @@ Hooks.on('renderSpellPointsForm', (spellPointsForm, html, data) => {
 
 Hooks.on("dnd5e.preItemUsageConsumption", (item, consume, options, update) => {
   SpellPoints.castSpell(item, consume, options, update);
+})
+
+Hooks.on("renderItemSheet", async (app, html, data) => {
+  SpellPoints.renderSpellPointsItem(app, html, data);
 })