+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 - 1.2.0 First support for FoundryVTT 0.8+
 - 1.2.1 Added Spanish translation.
 - 1.2.2 Display spell point cost in Usage Configuration window when casting a spell, fix a bug where the insufficient spell points warning would not pop up when casting
-- 1.2.3 Bugfix: https://github.com/misthero/dnd5e-spellpoints/issues/13.
-- 1.2.4 Feature: Disable Spell Point switch on character sheet when sheet is locked. https://github.com/misthero/dnd5e-spellpoints/issues/14
+- 1.2.3 Bugfix: <https://github.com/misthero/dnd5e-spellpoints/issues/13>.
+- 1.2.4 Feature: Disable Spell Point switch on character sheet when sheet is locked. <https://github.com/misthero/dnd5e-spellpoints/issues/14>
 - 1.2.5 Bugfix: Restored mixed mode checkbox on certain character sheets
 - 1.3.0 Add a new options to the formula selection box. The new options allow custom formulas defined at runtime for calculating spell point costs and maximum spell points. Custom formulas are hidden from configuration whenever a non-custom formula is selected. Selecting a formula will override options within the configuration window.
 - 1.3.1 Compatibility fixes for FVTT 0.9x
@@ -26,3 +26,9 @@
   - Fixed readme
   - Removed message to create resource if actor doesn't have spell point item.
   - New feature, enable for NPC (setting available to enable or disable this feature)
+- 2.2.0 New Features
+  - Per character/class spellpoints progression and cost.
+  - Per item spell points progression and cost.
+  - Customizable spell point bar color.
+  - New non-linear progression by level (supports roll formulas).
+  - Factor the progression based on caster type: full, half, third, pact, artificier.